5.x Make global functions opt-in
Remove the automatic loading of functions_global.php and add an opt-in
function loader. This is a breaking change from 4.x but we'll give
notice and tools to upgrade and its a one-line fix if you don't.

If you want to keep using global functions just add

```
require CAKE . 'functions.php';
```
to your bootstrap.php.